+++ packages/google_sign_in/test/google_sign_in_test.dart
@@ -16,10 +16,10 @@ void main() {
     );
 
     const Map<String, String> kUserData = <String, String>{
-      'email': 'john.doe@gmail.com',
-      'id': '8162538176523816253123',
-      'photoUrl': 'https://lh5.googleusercontent.com/photo.jpg',
-      'displayName': 'John Doe',
+      "email": "john.doe@gmail.com",
+      "id": "8162538176523816253123",
+      "photoUrl": "https://lh5.googleusercontent.com/photo.jpg",
+      "displayName": "John Doe",
     };
 
     const Map<String, dynamic> kDefaultResponses = <String, dynamic>{
@@ -40,12 +40,12 @@ void main() {
     GoogleSignIn googleSignIn;
 
     setUp(() {
-      responses = new Map<String, dynamic>.from(kDefaultResponses);
+      responses = Map<String, dynamic>.from(kDefaultResponses);
       channel.setMockMethodCallHandler((MethodCall methodCall) {
         log.add(methodCall);
-        return new Future<dynamic>.value(responses[methodCall.method]);
+        return Future<dynamic>.value(responses[methodCall.method]);
       });
-      googleSignIn = new GoogleSignIn();
+      googleSignIn = GoogleSignIn();
       log.clear();
     });
 
@@ -267,14 +267,14 @@ void main() {
 
     test('signInSilently suppresses errors by default', () async {
       channel.setMockMethodCallHandler((MethodCall methodCall) {
-        throw 'I am an error';
+        throw "I am an error";
       });
       expect(await googleSignIn.signInSilently(), isNull); // should not throw
     });
 
     test('signInSilently forwards errors', () async {
       channel.setMockMethodCallHandler((MethodCall methodCall) {
-        throw 'I am an error';
+        throw "I am an error";
       });
       expect(googleSignIn.signInSilently(suppressErrors: false),
           throwsA(isInstanceOf<PlatformException>()));
@@ -286,17 +286,17 @@ void main() {
         if (methodCall.method == 'init') {
           initCount++;
           if (initCount == 1) {
-            throw 'First init fails';
+            throw "First init fails";
           }
         }
-        return new Future<dynamic>.value(responses[methodCall.method]);
+        return Future<dynamic>.value(responses[methodCall.method]);
       });
       expect(googleSignIn.signIn(), throwsA(isInstanceOf<PlatformException>()));
       expect(await googleSignIn.signIn(), isNotNull);
     });
 
     test('created with standard factory uses correct options', () async {
-      googleSignIn = new GoogleSignIn.standard();
+      googleSignIn = GoogleSignIn.standard();
 
       await googleSignIn.signInSilently();
       expect(googleSignIn.currentUser, isNotNull);
@@ -315,7 +315,7 @@ void main() {
 
     test('created with defaultGamesSignIn factory uses correct options',
         () async {
-      googleSignIn = new GoogleSignIn.games();
+      googleSignIn = GoogleSignIn.games();
 
       await googleSignIn.signInSilently();
       expect(googleSignIn.currentUser, isNotNull);
@@ -332,79 +332,54 @@ void main() {
       );
     });
 
-    test('getTokens', () async {
+    test('getAuthentication', () async {
       await googleSignIn.signIn();
       log.clear();
 
       final GoogleSignInAccount user = googleSignIn.currentUser;
+      final GoogleSignInAuthentication auth = await user.getAuthentication();
 
-      final GoogleSignInAuthentication auth = await user.authentication;
+      expect(auth.accessToken, '456');
+      expect(auth.idToken, '123');
       expect(
         log,
         <Matcher>[
           isMethodCall('getTokens', arguments: <String, dynamic>{
             'email': 'john.doe@gmail.com',
+            'shouldRecoverAuth': true,
           }),
         ],
       );
-      expect(auth.accessToken, '456');
-      expect(auth.idToken, '123');
-    });
 
-    test('getTokens throws $UserRecoverableAuthException', () async {
-      await googleSignIn.signIn();
-      final GoogleSignInAccount user = googleSignIn.currentUser;
-
-      channel.setMockMethodCallHandler((MethodCall call) {
-        if (call.method == 'getTokens') {
-          throw new PlatformException(
-            code: 'user_recoverable_auth',
-            message: 'msg',
-          );
-        }
-
-        return null;
-      });
-
+      log.clear();
+      final GoogleSignInAuthentication authFalseRecover =
+          await user.getAuthentication(shouldRecoverAuth: false);
       expect(
-        () async => await user.authentication,
-        throwsA(isInstanceOf<UserRecoverableAuthException>()),
-      );
-    });
-
-    test('getTokens throws $PlatformException', () async {
-      await googleSignIn.signIn();
-      final GoogleSignInAccount user = googleSignIn.currentUser;
-
-      channel.setMockMethodCallHandler((MethodCall call) {
-        if (call.method == 'getTokens') {
-          throw new PlatformException(code: 'code', message: 'msg');
-        }
-
-        return null;
-      });
-
-      expect(
-        () async => await user.authentication,
-        throwsA(isInstanceOf<PlatformException>()),
+        log,
+        <Matcher>[
+          isMethodCall('getTokens', arguments: <String, dynamic>{
+            'email': 'john.doe@gmail.com',
+            'shouldRecoverAuth': false,
+          }),
+        ],
       );
     });
   });
 
   group('GoogleSignIn with fake backend', () {
     const FakeUser kUserData = FakeUser(
-      id: '8162538176523816253123',
-      displayName: 'John Doe',
-      email: 'john.doe@gmail.com',
-      photoUrl: 'https://lh5.googleusercontent.com/photo.jpg',
+      id: "8162538176523816253123",
+      displayName: "John Doe",
+      email: "john.doe@gmail.com",
+      photoUrl: "https://lh5.googleusercontent.com/photo.jpg",
     );
 
     GoogleSignIn googleSignIn;
 
     setUp(() {
       GoogleSignIn.channel.setMockMethodCallHandler(
-          (new FakeSignInBackend()..user = kUserData).handleMethodCall);
-      googleSignIn = new GoogleSignIn();
+          (FakeSignInBackend()..user = kUserData).handleMethodCall);
+      googleSignIn = GoogleSignIn();
     });
 
     test('user starts as null', () async {